requirements.txt
streamlit
pandas
numpy
faiss-cpu
sentence-transformers
transformers
torch
sentencepiece